requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
crispy-bootstrap5==2023.10
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-braces==1.15.0
django-crispy-forms==2.1
django-import-export==3.3.7
django-nested-admin==4.0.2
et-xmlfile==1.1.0
MarkupPy==1.14
mysqlclient==2.2.0
odfpy==1.4.1
openpyxl==3.1.2
package-name==0.1
pillow==10.2.0
psycopg2==2.9.9
python-decouple==3.8
python-monkey-business==1.0.0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
